refactor(service): split the middleware path out of forward.rs
Address the review: forward.rs (1033 lines) carried two responsibilities — the
direct buffered/streaming forward hot paths, and the entire optional
UDS-middleware split sub-protocol. Extract the latter
(forward_chat_completion_for_middleware, build_middleware_receipt_prefix, and the
three finalize_middleware_* methods) into a new sibling `middleware.rs` as its
own `impl AciService` block.

forward.rs (531) now owns the direct paths + shared upstream-verification/receipt
helpers; middleware.rs (520) owns the middleware forward + finalization. The
shared helpers stay pub(super) at the service level, so claims.rs and the direct
paths still reach them. No behavior change; cargo test --all-targets: 220 passed.
